+++ src/Element/Checkbox.php
@@ -26,11 +26,12 @@ class Checkbox extends BaseInput {
         $this->element = array_merge([
             "class" => "js-icheck",
             "name" => $this->name,
-            "id" => "field_" . $this->name
+            "id" => "field_" . $this->name,
+            "binary" => true
         ], $this->element);
 
         // We add the check status instead of the value
-        if ($this->getValue() == 1) {
+        if ($this->element["binary"] !== false && $this->getValue() == 1) {
             $this->element["checked"] = "checked";
         }
     }