+++ CHANGELOG.md
@@ -1,5 +1,35 @@
 # Change Log
 
+## 2.2.9
+- Fix issue when setting data that is a collection
+- `formSuccess` and `confirmSuccess` events now include the request data as a second argument
+
+## 2.2.8
+- Fix icon in textarea macro
+
+## 2.2.7
+- Added `modal-large` template file.
+
+## 2.2.6
+- Fix issue with `binary` checkbox tests.
+- Fix Text input style when no icon is added
+
+## 2.2.5
+- Added `binary` option for checkbox to disable UF binary checkbox system (bool; default true).
+
+## 2.2.4
+- Add necessary HTML to disable submit and cancel button in modal form.
+
+## 2.2.3
+- New `$form->setOptions` function to set options of a select element. Shortcut for using `setInputArgument` and `setValue`.
+
+## 2.2.2
+- Fix issue with error alert no displaying on confirmation dialog
+
+## 2.2.1
+- Initialize ufAlert if not already done
+- Autofocus first form field when modal is displayed
+
 ## 2.2.0
 - Refactored the javascript plugin
 - Added new events
@@ -10,11 +40,11 @@
 
 ## 2.1.1
 - Fix issue with the select macro
-- Renamed macro templates with the `*.html.twig` extension 
+- Renamed macro templates with the `*.html.twig` extension
 
 ## 2.1.0
 - Completely refactored how form fields are parsed, including how default value are defined. Each input type now defines it's own class for defining default values and transforming some input.
-- Twig templates updated to reflect the new parser. 
+- Twig templates updated to reflect the new parser.
 - Twig macros changed from `*.generate(name, value)` to `*.generate(input)`.
 - **`Bool` type changed to `checkbox`**.
 - Removed the `number` Twig template (Will use the text input one).